+++ pyslvs_ui/widgets/main_abc.py
@@ -71,7 +71,7 @@ class MainWindowABC(QMainWindow, Ui_MainWindow, metaclass=QABCMeta):
         raise NotImplementedError
 
     @abstractmethod
-    def delete_empty_links(self) -> None:
+    def delete_redundant_links(self) -> None:
         raise NotImplementedError
 
     @abstractmethod